profiles: curl: switch to user-tmp abstraction
The curl profile allows reading and writing to /tmp/ so instead of
two rules that don't cover all tmp locations, switch to the user-tmp
abstraction to allow access to the various possible tmp locations.

Note: The does reduce the write permission to owner write, instead
of the wider file w /tmp/**,

In addition move the @{HOME} permissions to be restricted files owned
by the user.

Signed-off-by: John Johansen <john.johansen@canonical.com>