regression: avoid exposing disconnected_mount_complain's mounts externally
Due to a behavioral change between 6.14 and 6.15, the move_mount that we
set up in disconnected_mount_complain's open_tree test would now persist,
causing issues with test cleanup after that test. Ensure that we remount
all the mounts as private after unsharing the mount namespace to prevent
this from happening.

Signed-off-by: Ryan Lee <ryan.lee@canonical.com>